hwc: Fix unnecessary pipe allocation from mdp comp.
In High res case, there was no check for whether a layer is marked for mdp
leading to unnecessary pipes reserved for rgb layers on external.

Remove z order setting in allocLayerPipes, its too early, incorrect and
will be over-written anyway.

Remove the separate pipe allocation block for YUV. It is redundant and
unnecessary.

Change-Id: Ie90b4202a27ca4f52874514e3fa2b136ae5717f5